Refactor tests
  - make fusionbox.core.tests a package
  - move unrelated tests into their own files
  - run fusionbox.core.tests in test_app.tests
