+++ README.md
@@ -15,3 +15,12 @@ Contributions to `django-fusionbox` are welcome provided they adhere to the foll
 ## Documentation
 
 http://django-fusionbox.readthedocs.org/
+
+## Running the tests
+`django-fusionbox` comes with a test project to run its tests.
+
+    mktmpenv
+    python setup.py develop
+    cd test_project
+    pip install -r requirements.txt
+    ./manage.py test test_app
