+++ fusionbox/core/tests/__init__.py
@@ -1,0 +1,2 @@
+from test_templatetags import *
+from test_behaviors import *
