+++ test_project/test_app/tests.py
@@ -1,0 +1,1 @@
+from fusionbox.core.tests import *
